pyproject.toml
# MIT License
#
# Copyright (C) 2024 The Little Hackers.
#
# Permission is hereby granted, free of charge, to any person obtaining
# a copy of this software and associated documentation files (the
# "Software"), to deal in the Software without restriction, including
# without limitation the rights to use, copy, modify, merge, publish,
# distribute, sublicense, and/or sell copies of the Software, and to
# permit persons to whom the Software is furnished to do so, subject to
# the following conditions:
#
# The above copyright notice and this permission notice shall be
# included in all copies or substantial portions of the Software.
#
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
# MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
# IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
# CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
# TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
# SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

[tool.poetry]
authors = ["Daniel CAUNE <daniel.caune@thelittlehackers.com>"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries",
]
description = "Python utility library for manipulating images"
keywords = ["image", "library", "utils"]
license = "MIT License"
name = "thelittlehackers-image-utility"
packages = [{include = "thelittlehackers", from = "src"}]
readme = "README.md"
repository = "https://github.com/the-little-hackers/library_image-utility_python"
version = "0.0.2"

[tool.poetry.dependencies]
pillow = "^11.2.1"
python = ">=3.12,<4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
